refactor: reorganize :feature:chat by sub-features (chat + search)
Move all files from flat com.bina.chat.* packages into two explicit
sub-feature packages:
- com.bina.chat.chat.*  — conversation, ViewModel, domain, presentation
- com.bina.chat.search.* — CharacterSearchApiService, DataSource, models

DI stays at com.bina.chat.di and wires both sub-features together.
