+++ feature/chat/src/main/java/com/bina/chat/chat/data/model/ChatMessageData.kt
@@ -1,0 +1,7 @@
+package com.bina.chat.chat.data.model
+
+data class ChatMessageData(
+    val role: String,
+    val text: String,
+    val timestampMs: Long
+)
