+++ feature/chat/src/main/java/com/bina/chat/chat/data/model/ToolResponse.kt
@@ -1,0 +1,11 @@
+package com.bina.chat.chat.data.model
+
+data class ToolResponse(
+    val text: String?,
+    val functionCalls: List<FunctionCallData>
+)
+
+data class FunctionCallData(
+    val name: String,
+    val args: Map<String, String?>
+)
