+++ feature/chat/src/main/java/com/bina/chat/chat/domain/model/AgentMessageResult.kt
@@ -1,0 +1,6 @@
+package com.bina.chat.chat.domain.model
+
+data class AgentMessageResult(
+    val text: String,
+    val navigationEvent: ChatNavigationEvent? = null
+)
