+++ feature/chat/src/main/java/com/bina/chat/chat/domain/model/ChatMessageDomain.kt
@@ -1,0 +1,7 @@
+package com.bina.chat.chat.domain.model
+
+data class ChatMessageDomain(
+    val role: MessageRole,
+    val text: String,
+    val timestampMs: Long
+)
