+++ feature/chat/src/main/java/com/bina/chat/chat/domain/model/MessageRole.kt
@@ -1,0 +1,3 @@
+package com.bina.chat.chat.domain.model
+
+enum class MessageRole { USER, AI }
