+++ feature/chat/src/main/java/com/bina/chat/chat/domain/model/ModelAvailability.kt
@@ -1,0 +1,7 @@
+package com.bina.chat.chat.domain.model
+
+sealed class ModelAvailability {
+    object Available : ModelAvailability()
+    object Unavailable : ModelAvailability()
+    object Downloadable : ModelAvailability()
+}
