+++ feature/chat/src/main/java/com/bina/chat/chat/domain/repository/ChatRepository.kt
@@ -1,0 +1,13 @@
+package com.bina.chat.chat.domain.repository
+
+import com.bina.chat.chat.domain.model.AgentMessageResult
+import com.bina.chat.chat.domain.model.ModelAvailability
+import kotlinx.coroutines.flow.Flow
+
+interface ChatRepository {
+    suspend fun checkAvailability(): ModelAvailability
+    suspend fun warmup()
+    fun streamResponse(userMessage: String): Flow<String>
+    suspend fun sendAgentMessage(userMessage: String): AgentMessageResult
+    fun close()
+}
