--- feature/chat/src/main/java/com/bina/chat/data/datasource/CharacterSearchDataSource.kt
@@ -1,5 +1,0 @@
-package com.bina.chat.data.datasource
-
-interface CharacterSearchDataSource {
-    suspend fun searchByName(name: String): Int?
-}
