--- feature/chat/src/main/java/com/bina/chat/data/datasource/CharacterSearchDataSourceImpl.kt
@@ -1,11 +1,0 @@
-package com.bina.chat.data.datasource
-
-import com.bina.chat.data.remote.CharacterSearchApiService
-
-class CharacterSearchDataSourceImpl(
-    private val api: CharacterSearchApiService
-) : CharacterSearchDataSource {
-
-    override suspend fun searchByName(name: String): Int? =
-        runCatching { api.searchByName(name).results?.firstOrNull()?.id }.getOrNull()
-}
