--- feature/chat/src/main/java/com/bina/chat/data/model/CharacterSearchResponse.kt
@@ -1,10 +1,0 @@
-package com.bina.chat.data.model
-
-data class CharacterSearchResponse(
-    val results: List<CharacterSearchResult>?
-)
-
-data class CharacterSearchResult(
-    val id: Int,
-    val name: String
-)
