--- feature/chat/src/main/java/com/bina/chat/data/model/ChatMessageData.kt
@@ -1,7 +1,0 @@
-package com.bina.chat.data.model
-
-data class ChatMessageData(
-    val role: String,
-    val text: String,
-    val timestampMs: Long
-)
