--- feature/chat/src/main/java/com/bina/chat/data/remote/CharacterSearchApiService.kt
@@ -1,10 +1,0 @@
-package com.bina.chat.data.remote
-
-import com.bina.chat.data.model.CharacterSearchResponse
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface CharacterSearchApiService {
-    @GET("character")
-    suspend fun searchByName(@Query("name") name: String): CharacterSearchResponse
-}
