--- feature/chat/src/main/java/com/bina/chat/domain/model/AgentMessageResult.kt
@@ -1,6 +1,0 @@
-package com.bina.chat.domain.model
-
-data class AgentMessageResult(
-    val text: String,
-    val navigationEvent: ChatNavigationEvent? = null
-)
