--- feature/chat/src/main/java/com/bina/chat/domain/model/ChatMessageDomain.kt
@@ -1,7 +1,0 @@
-package com.bina.chat.domain.model
-
-data class ChatMessageDomain(
-    val role: MessageRole,
-    val text: String,
-    val timestampMs: Long
-)
