--- feature/chat/src/main/java/com/bina/chat/domain/model/ChatNavigationEvent.kt
@@ -1,6 +1,0 @@
-package com.bina.chat.domain.model
-
-sealed class ChatNavigationEvent {
-    data class OpenCharacter(val characterId: Int) : ChatNavigationEvent()
-    data class SearchCharacters(val query: String) : ChatNavigationEvent()
-}
