--- feature/chat/src/main/java/com/bina/chat/domain/model/MessageRole.kt
@@ -1,3 +1,0 @@
-package com.bina.chat.domain.model
-
-enum class MessageRole { USER, AI }
