--- feature/chat/src/main/java/com/bina/chat/domain/model/ModelAvailability.kt
@@ -1,7 +1,0 @@
-package com.bina.chat.domain.model
-
-sealed class ModelAvailability {
-    object Available : ModelAvailability()
-    object Unavailable : ModelAvailability()
-    object Downloadable : ModelAvailability()
-}
