--- feature/chat/src/main/java/com/bina/chat/domain/usecase/SendMessageUseCase.kt
@@ -1,9 +1,0 @@
-package com.bina.chat.domain.usecase
-
-import com.bina.chat.domain.model.AgentMessageResult
-import com.bina.chat.domain.repository.ChatRepository
-
-class SendMessageUseCase(private val repository: ChatRepository) {
-    suspend operator fun invoke(userMessage: String): AgentMessageResult =
-        repository.sendAgentMessage(userMessage)
-}
