--- feature/chat/src/main/java/com/bina/chat/presentation/mapper/ChatMessageUiMapper.kt
@@ -1,11 +1,0 @@
-package com.bina.chat.presentation.mapper
-
-import com.bina.chat.domain.model.ChatMessageDomain
-import com.bina.chat.presentation.model.ChatMessageUiModel
-
-class ChatMessageUiMapper {
-    fun map(domain: ChatMessageDomain): ChatMessageUiModel = ChatMessageUiModel(
-        role = domain.role,
-        text = domain.text
-    )
-}
