--- feature/chat/src/main/java/com/bina/chat/presentation/model/ChatMessageUiModel.kt
@@ -1,9 +1,0 @@
-package com.bina.chat.presentation.model
-
-import com.bina.chat.domain.model.MessageRole
-
-data class ChatMessageUiModel(
-    val role: MessageRole,
-    val text: String,
-    val isStreaming: Boolean = false
-)
