--- feature/chat/src/main/java/com/bina/chat/presentation/state/ChatUiState.kt
@@ -1,14 +1,0 @@
-package com.bina.chat.presentation.state
-
-import com.bina.chat.presentation.model.ChatMessageUiModel
-
-sealed class ChatUiState {
-    object Initializing : ChatUiState()
-    object ModelUnavailable : ChatUiState()
-    object ModelDownloadable : ChatUiState()
-    data class Conversation(
-        val messages: List<ChatMessageUiModel>,
-        val isAiTyping: Boolean,
-        val errorMessage: String? = null
-    ) : ChatUiState()
-}
