+++ feature/chat/src/main/java/com/bina/chat/search/data/datasource/CharacterSearchDataSource.kt
@@ -1,0 +1,5 @@
+package com.bina.chat.search.data.datasource
+
+interface CharacterSearchDataSource {
+    suspend fun searchByName(name: String): Int?
+}
