+++ feature/chat/src/main/java/com/bina/chat/search/data/datasource/CharacterSearchDataSourceImpl.kt
@@ -1,0 +1,11 @@
+package com.bina.chat.search.data.datasource
+
+import com.bina.chat.search.data.remote.CharacterSearchApiService
+
+class CharacterSearchDataSourceImpl(
+    private val api: CharacterSearchApiService
+) : CharacterSearchDataSource {
+
+    override suspend fun searchByName(name: String): Int? =
+        runCatching { api.searchByName(name).results?.firstOrNull()?.id }.getOrNull()
+}
