+++ feature/chat/src/main/java/com/bina/chat/search/data/model/CharacterSearchResponse.kt
@@ -1,0 +1,10 @@
+package com.bina.chat.search.data.model
+
+data class CharacterSearchResponse(
+    val results: List<CharacterSearchResult>?
+)
+
+data class CharacterSearchResult(
+    val id: Int,
+    val name: String
+)
