+++ feature/chat/src/main/java/com/bina/chat/search/data/remote/CharacterSearchApiService.kt
@@ -1,0 +1,10 @@
+package com.bina.chat.search.data.remote
+
+import com.bina.chat.search.data.model.CharacterSearchResponse
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface CharacterSearchApiService {
+    @GET("character")
+    suspend fun searchByName(@Query("name") name: String): CharacterSearchResponse
+}
